+++ static/js/set-password.js
@@ -1,0 +1,8 @@
+$(function () {
+    $pw = $("#password");
+    $meter = $("#meter");
+    $pw.on("input", function() {
+        var result = zxcvbn($pw.val());
+        $meter.attr("class", "score-" + result.score);
+    });
+});