+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Can configure the email integration to only report the "down" events (#231)
 - Add "Test!" function in the Integrations page (#207)
 - Rate limiting for the log in attempts
+- Password strength meter and length check in the "Set Password" form
 
 ## 1.6.0 - 2019-04-01